% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExampleSet.R
\name{GetExampleSet}
\alias{GetExampleSet}
\title{Get example dataset}
\usage{
GetExampleSet(exampleData)
}
\arguments{
\item{exampleData}{A character, should be one of"GEP","clinicaldata",
"TMEcellinfo", "TME_related_Goterm", "Jaccardscore" and "GoCellconGene".}
}
\value{
example dataset
}
\description{
This function is used to achieve example dataset.
}
