% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iTMEcell.R
\name{iTMEcell}
\alias{iTMEcell}
\title{Identification of abnormal tumor microenvironment (TME) cells}
\usage{
iTMEcell(ExpData,Condition.label,nperm=1000)
}
\arguments{
\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{Condition.label}{A data.frame at least two columns which are “sample” (sample.id) and ”Condition.label” (condition of samples, “0” represents case and “1” represents control).}

\item{nperm}{Number of random permutations (default: 1000).}
}
\value{
A dataframe with seven columns those are cell names, marker source, marker size, marker genes, centrality (eigenvector centrality), P-value and FDR.
}
\description{
The function "iTMEcell" is used to calculate the eigenvector centrality of TME cells and identify abnormal TME cells.
}
\examples{
library(igraph)
#Obtain input data
GEP<-GetExampleSet('GEP')
Condition.label<-GetExampleSet('Condition.label')
#Run the function
iTMEcellresult<-iTMEcell(ExpData=GEP,Condition.label=Condition.label,nperm=1000)
}
