% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforest.R
\name{plotforest}
\alias{plotforest}
\title{Draw a forest plot.}
\usage{
plotforest(Regress.list,p.cutoff=0.05,g.pos=2,b.size=3,
        col=c("#FE0101","#1C61B6","#A4A4A4"),
        lwd.zero=2,lwd.ci=3,x.lab="Hazard Ratio Plot")
}
\arguments{
\item{Regress.list}{The result of function "RiskRegressModel".}

\item{p.cutoff}{Statistical significance threshold of cox regression analysis, based on which to determine the genes used to draw the graph.}

\item{g.pos}{A number to control the position of the graph element in forestplot.}

\item{b.size}{A number to control the box size.}

\item{col}{Vector of colors including three color code which are corresponding to box, box line and reference line.}

\item{lwd.zero}{A number to control the thickness of the reference line.}

\item{lwd.ci}{A number to control the thickness of the box line.}

\item{x.lab}{Setting the title.}
}
\value{
A forest plot
}
\description{
The function "plotforest" is used to draw a forest plot according to the result of cox analysis from function "RiskRegressModel".
}
\examples{
library(forestplot)
library(survival)
#Obtain input data
GEP<-GetExampleSet('GEP')
clinicaldata<-GetExampleSet('clinicaldata')
#Run the function
R.result<-RiskRegressModel(cellname='NK cells',ExpData=GEP,clinical=clinicaldata,
     p.cutoff=0.05)
plotforest(Regress.list=R.result,p.cutoff=0.05)
}
