% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Draw a heat map.}
\usage{
plotHeatmap(Regress.list,ExpData,cut.off=NULL,p.cutoff=0.05,cluster.rows=F,
cluster.cols=F,bk=c(-2.4,2.3),show.rownames=T,show.colnames=F,
ann_colors=c("#FFAA2C","#2CBADA"),col=c("#2A95FF","#FF1C1C"))
}
\arguments{
\item{Regress.list}{The result of function "RiskRegressModel".}

\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{cut.off}{Samples will be grouped according to this threshold. If not specified, the sample will be grouped according to the median risk score.}

\item{p.cutoff}{Statistical significance threshold of cox regression analysis, based on which to determine the genes used to draw the graph.}

\item{cluster.rows}{Boolean values determining if rows should be clustered or hclust object.}

\item{cluster.cols}{Boolean values determining if columns should be clustered or hclust object.}

\item{bk}{A numeric vector that covers the range of values. Users could adjust color depth through this parameter.}

\item{show.rownames}{Boolean specifying if row names are be shown.}

\item{show.colnames}{Boolean specifying if column names are be shown.}

\item{ann_colors}{Vector of colors for specifying the color of column annotation.}

\item{col}{Vector of colors used in heat map.}
}
\value{
A heat map
}
\description{
The function "plotHeatmap" is used to draw a heat map of marker genes.
}
\examples{
library(pheatmap)
library(survival)
#Obtain input data
GEP<-GetExampleSet('GEP')
clinicaldata<-GetExampleSet('clinicaldata')
#Run the function
R.result<-RiskRegressModel(cellname='NK cells',ExpData=GEP,clinical=clinicaldata,
      p.cutoff=0.05)
plotHeatmap(Regress.list=R.result,ExpData=GEP,p.cutoff=0.05)
}
