% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKMcurve.R
\name{plotKMcurve}
\alias{plotKMcurve}
\title{Draw a Kaplan-Meier curve.}
\usage{
plotKMcurve(Regress.list,ExpData,risk.table=TRUE,labs=c("High risk","Low risk"),
                   title="Group",line.col=c("#FFAA2C", "#2CBADA"))
}
\arguments{
\item{Regress.list}{The result of function "RiskRegressModel".}

\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{risk.table}{TRUE or FALSE specifying whether to show or not the risk table. Default is TRUE.}

\item{labs}{A character vector for specifying legend labels.}

\item{title}{legend title}

\item{line.col}{Vector of colors for specifying the color of curve.}
}
\value{
Kaplan-Meier curve
}
\description{
The function "plotKMcurve" is used to draw the Kaplan-Meier curve according to the riskscore of samples from function "RiskRegressModel".
}
\examples{
library(survival)
library(survminer)
#Obtain input data
GEP<-GetExampleSet('GEP')
clinicaldata<-GetExampleSet('clinicaldata')
#Run the function
R.result<-RiskRegressModel(cellname='NK cells',ExpData=GEP,clinical=clinicaldata,
       p.cutoff=0.05)
plotKMcurve(Regress.list=R.result,ExpData=GEP)
}
