% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSplitViolin.R
\name{plotSplitViolin}
\alias{plotSplitViolin}
\title{Draw a split violin plot.}
\usage{
plotSplitViolin(Regress.list,ExpData,gene.name,method="t.test",
compare.label="p.signif",col=c("#E69F00", "#56B4E9"),x.ceiling=15,
y.lab="Gene Expression",x.lab=NULL,title=NULL)
}
\arguments{
\item{Regress.list}{The result of function "RiskRegressModel".}

\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{gene.name}{A gene symbol in inputted gene expression profile.}

\item{method}{A character string indicating which method to be used for comparing means. The default method is "t.test". Other three methods are "wilcox.test", "anova" and "kruskal.test".}

\item{compare.label}{A character string specifying label type. Allowed values include "p.signif" (shows the significance levels), "p.format" (shows the formatted p value).}

\item{col}{Vector of colors used to specify the color of different groups.}

\item{x.ceiling}{The maximum value of the y axis.}

\item{y.lab}{Setting the title of the y-axis.}

\item{x.lab}{Setting the title of the x-axis.}

\item{title}{Setting the title}
}
\value{
A split violin plot
}
\description{
The function "plotSplitViolin" is used to draw a split violin plot of gene expression.
}
\examples{
library(ggplot2)
library(reshape2)
library(plyr)
library(ggpubr)
library(survival)
#Obtain input data
GEP<-GetExampleSet('GEP')
clinicaldata<-GetExampleSet('clinicaldata')
#Run the function
R.result<-RiskRegressModel(cellname='NK cells',ExpData=GEP,clinical=clinicaldata,
        p.cutoff=0.05)
plotSplitViolin(Regress.list=R.result,ExpData=GEP,gene.name="PDCD1")
}
