% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskRegressModel.R
\name{RiskRegressModel}
\alias{RiskRegressModel}
\title{Constructing the cox risk regression model with cell's marker genes}
\usage{
RiskRegressModel(cellname,ExpData,clinical,marker=NULL,method = "lasso",p.cutoff=0.05)
}
\arguments{
\item{cellname}{A cell whose marker genes will be used to perform regression analysis. The format of the entered cell name should refer to the cell information we provide.}

\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{clinical}{A dataframe with three columns which are "sample" (sample id),"status" (survival status of samples, "0" represents live and "1" represents dead) and "time" (survival time of samples).}

\item{marker}{A character vector composed with marker genes. If you does not want to use the marker genes provided by us, you can specify the marker genes you need with this parameter.}

\item{method}{This parameter specifies the method of regression analysis. "method=cox": Only univariate regression analysis was performed, and the model was constructed with coefficients. "method=lasso"(default): The significant variables in univariate analysis were used for LASSO regression analysis, and the coefficients of LASSO analysis were used to construct the model.}

\item{p.cutoff}{Statistical significance threshold for regression analysis (default: 0.05).}
}
\value{
A list with two dataframes which are riskscores of samples and result of cox regression analysis respectively.
}
\description{
This function is used to perform regression analysis and build risk regression models.
}
\details{
In the default method, users can specify a cell, and then the function will perform cox regression analysis on expression value of marker genes of the cell and survival data. Statistical significant genes will be selected for further lasso regression analysis. Finally, the lasso regression coefficients were used to weight the gene expression values to calculate the risk score for samples.
}
\examples{
library(survival)
#Obtain input data
GEP<-GetExampleSet('GEP')
clinicaldata<-GetExampleSet('clinicaldata')
#Run the function
R.result<-RiskRegressModel(cellname='NK cells',ExpData=GEP,
  clinical=clinicaldata,method = "lasso",p.cutoff=0.05)
}
