% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{envData}
\alias{envData}
\title{An environment variable which includes some example data}
\format{
An environment variable
}
\usage{
envData
}
\description{
An environment variable which includes some example data.
iTMEcellresult: The result of function "iTMEcell".
GEP: An example gene expression profile.
TME_related_Goterm: Biological Process data from Gene Ontology. Goterms associated with TME were selected through hypergeometric test.
GoCellconGene: Genes symble shared by a pair of cell with Biological Process.
Jaccardscore: Jaccard score shared by a pair of cell with Biological Process.
clinicaldata: Clinical data of samples in gene expression profile.
TMEcellinfo: TME cells information.
}
\keyword{datasets}
