% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskRegressModel.R
\name{RiskRegressModel}
\alias{RiskRegressModel}
\title{Constructing the cox risk regression model with cell marker genes}
\usage{
RiskRegressModel(cellname,ExpData,clinical,marker=NULL,p.cutoff=0.01,method='univ')
}
\arguments{
\item{cellname}{A cell whose marker genes will be used to perform regression analysis. The format of the entered cell name should refer to the cell information we provide.}

\item{ExpData}{A gene expression profile of interest (rows are genes, columns are samples).}

\item{clinical}{A dataframe with three columns which are "sample" (sample id),"status" (survival status of samples, "0" represents live and "1" represents dead) and "time" (survival time of samples).}

\item{marker}{A character vector composed with marker genes. If you does not want to use the marker genes provided by us, you can specify the marker genes you need with this parameter.}

\item{p.cutoff}{Statistical significance threshold for regression analysis (default: 0.05).}

\item{method}{Users can specify the univariate cox regression analysis (default, method = "univ") or multivariate cox regression analysis (method = "muti") in this function.}
}
\value{
A list with two dataframes which are riskscores of samples and result of cox regression analysis respectively.
}
\description{
Users can specify a cell, and then the function will perform cox regression analysis on expression of marker genes of the cell and survival data. Statistical significant genes will be selected, and with these genes, a gene risk score model was constructed using a formula derived from the expression of the genes weighted by their cox proportional hazards regression coefficient.
}
\examples{
library(survival)
#Obtain input data
GEP<-GetExampleSet('GEP')
clinicaldata<-GetExampleSet('clinicaldata')
#Run the function
R.result<-RiskRegressModel(cellname='NK cells',ExpData=GEP,
  clinical=clinicaldata,p.cutoff=0.01,method = 'univ')
}
